+++ tests/TestHarness.php
@@ -524,6 +524,7 @@ namespace GlpiPlugin\Samlsso\Config {
             public const CREATE_DATE     = 'date_creation';
             public const MOD_DATE        = 'date_mod';
             public const SYNC_ON_LOGIN   = 'sync_on_login';
+            public const REQUEST_TIMEOUT = 'request_timeout';
             public const SECURITY_WANTMESSAGESSIGNED = 'security_wantmessagessigned';
             public const SECURITY_WANTASSERTIONSSIGNED = 'security_wantassertionssigned';
             public const SECURITY_WANTASSERTIONSENCRYPTED = 'security_wantassertionsencrypted';